### 2025-06-25 #8482 @editorialbot
[PRE REVIEW]: PyMetric: A Geometry Informed Array Mathematics Package
**Submitting author:** @eliza-diggins (<a href="http://orcid.org/0009-0005-9389-9098">Eliza Diggins</a>)
**Repository:** https://github.com/Pisces-Project/PyMetric
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.2a
