### 2025-06-25 #8482 @danielskatz
@Eliza-Diggins - thanks for your submission.  I'll look for an editor next.

While I do so, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
