### 2025-08-13 #8482 @drvinceknight
@vissarion @anthbapt @jcorbino I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “PyMetric: A Geometry Informed Array Mathematics Package”. You can find more information at the top of this Github issue.
