### 2025-09-01 #8901 @editorialbot
[REVIEW]: PyMetric: A Geometry Informed Array Mathematics Package
**Submitting author:** @eliza-diggins (<a href="http://orcid.org/0009-0005-9389-9098">Eliza Diggins</a>)
**Repository:** https://github.com/Pisces-Project/PyMetric
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.3a1
