### 2025-09-04 #8901 @anthbapt
The quality of the writing can be improved or at least checked, some LLM artefacts "—" are present in the whole paper. For instance: "These systems can be highly nontrivial—such as ellipsoidal (homoeoidal) or spheroidal
15 coordinates—where explicitly" in the first paragraph of the summary.
