### 2025-09-04 #8901 @anthbapt
I recommend this paper. The codes, documentation, and accompanying paper provide insightful tools for applying differential geometry in a physical context, particularly by simplifying basic geometric operations.
The only limitation lies in the writing, which could benefit from some polishing, as I mentioned in my previous point.
