### 2025-10-18 f6bff81
Fixed Einsteinpy URL typo.
