+++ paper/paper.bib
@@ -17,7 +17,7 @@ url = {https://github.com/HDFGroup/hdf5}
 title = {EinsteinPy: Python library for General Relativity},
 author = {{EinsteinPy Development Team}},
 year = {2024},
-url = {https:/einsteinpy.org/},
+url = {https://einsteinpy.org/},
 }
 @article{perret2016dice,
   title={DICE: Disk Initial Conditions Environment},
