### 2025-10-18 2a4a0d5
Added DOI for YT reference.
