+++ paper/paper.bib
@@ -41,7 +41,8 @@ url = {https://einsteinpy.org/},
   number={1},
   pages={9},
   year={2010},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/0067-0049/192/1/9}
 }
 @article{sympy,
  title = {SymPy: symbolic computing in Python},
