### 2025-10-18 e4b5dd8
Revisions to paper + example in the paper.
