+++ paper/paper.md
@@ -25,38 +25,38 @@ bibliography: paper.bib
 PyMetric is a lightweight Python library designed to streamline differential geometry
 and vector calculus operations in user-defined coordinate systems, with a focus on applications
 in astrophysics and computational physics. The library was originally created to provide a
-geometric backend for the Pisces project, an (in development) general purpose astrophysical modeling and initial conditions
+geometric backend for the Pisces project, an in-development, general purpose astrophysical modeling and initial conditions
 library, but has since grown into an independent library due to its size and complexity. In many physical modeling tasks, it is both natural
 and advantageous to work in non-Cartesian coordinate systems that align with the inherent
-symmetries of the system. These systems can be highly nontrivial—such as ellipsoidal
-(homoeoidal) or spheroidal coordinates—where explicitly handling coordinate-specific
-expressions becomes tedious and error-prone. PyMetric provides a unified abstraction
+symmetries of the system. These coordinate systems can feature complex geometric structure which makes the
+explicit handling of differential operations cumbersome. This is particularly true for exotic coordinate
+systems (e.g. homoeoidal coordinate systems). PyMetric provides a unified abstraction
 that decouples the underlying coordinate representation from the operations themselves,
 allowing users to accurately compute gradients, divergences, Laplacians, and related geometric
-quantities through a consistent interface. This makes it easier to prototype and scale
+quantities through a consistent (and coordinate system agnostic) interface. This makes it easier to prototype and scale
 models in complex geometries without having to rewrite operations for each coordinate system.
 
 The core design of PyMetric relies on a hybrid symbolic-numeric model that balances efficiency,
 flexibility, and accuracy. Symbolic computation is used to derive key geometric quantities, such
-as metric tensors, Christoffel symbols, and Jacobians, directly from the structure of the coordinate
-system. These symbolic expressions preserve the full geometric context and can be reused across
-multiple evaluations. Once derived, they are compiled into optimized numerical functions that
-can be efficiently applied to array data on structured grids. This approach allows PyMetric to
-support coordinate-aware computation with minimal overhead, avoiding the need for repeated symbolic
-manipulation during runtime, while maintaining high accuracy through analytically correct geometric
+as metric tensors, Christoffel symbols, and Jacobians, from a minimal set of coordinate system properties.
+Once generated, these symbolic structures can be converted into efficient numerical routines that operate
+on array-backed data and are composed to perform higher-level operations.
+This approach allows PyMetric to support coordinate-aware computation with minimal overhead, avoiding the need for
+repeated symbolic manipulation during runtime, while maintaining high accuracy through analytically correct geometric
 expressions. The result is a powerful and extensible framework that enables NumPy-style [@harris2020array] workflows
 in complex coordinate systems without sacrificing physical fidelity.
 
 In addition to its symbolic-numeric foundation, PyMetric provides structured abstractions for grids
-and field data, supporting a range of coordinate systems and buffer backends—including in-memory arrays and
-HDF5 [@hdf5] storage for scalable computation. Users can define fields over geometric grids and apply
-differential operators without manually managing coordinate-dependent logic.
+and field data, supporting a range of coordinate systems and buffer backends. Because the PyMetric field abstraction
+is only minimally coupled to the underlying data storage, it can interface with a variety of array backends, including
+in-memory arrays and HDF5 [@hdf5] storage for lazy-loading and chunked computation. This design enables coordinate-aware
[…]
