### 2025-10-18 941cd0e
Improving figure caption.
