### 2025-10-30 #8901 @jcorbino
I have completed my review and recommend acceptance of this paper provided the following issues are addressed:

- The file PyMetric/docs/index.html redirects to https://eliza-diggins.github.io/Pisces-Geometry/build/html which cannot be found.
- In the **"Fields and Grids Operations"** section, briefly mention the types of grids supported or note any major limitations.
- Please correct the formatting of the list in the **"Future Development"** section.
