### 2025-11-16 15e524a
Paper revisions.
