+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={7825},
   pages={357--362},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-020-2649-2}
 }
 @software{hdf5,
 author = {{The HDF Group}},
@@ -18,6 +19,7 @@ title = {EinsteinPy: Python library for General Relativity},
 author = {{EinsteinPy Development Team}},
 year = {2024},
 url = {https://einsteinpy.org/},
+doi={10.48550/arXiv.2005.11288}
 }
 @article{perret2016dice,
   title={DICE: Disk Initial Conditions Environment},
@@ -31,7 +33,8 @@ url = {https://einsteinpy.org/},
   author={Yurin, Denis and Springel, Volker},
   journal={Astrophysics Source Code Library},
   pages={ascl--1408},
-  year={2014}
+  year={2014},
+  doi={10.48550/arXiv.1402.1623}
 }
 @article{turk2010yt,
   title={yt: A multi-code analysis toolkit for astrophysical simulation data},
