### 2026-01-14 #8901 @danielskatz
Congratulations to @Eliza-Diggins (Eliza Diggins) and co-author on your JOSS publication!!

And thanks to @jcorbino & @anthbapt for reviewing, and to @drvinceknight for editing
JOSS completely depends on volunteers and we couldn't do this without you.
